Spremembe narejene, deluje.
