+++ data/.app.json
@@ -23,10 +23,6 @@
     "velikost": 30,
     "tip": "freesansbold.ttf"
   },
-  "smer": {
-    "levo": "levo",
-    "desno": "desno"
-  },
   "slike": {
     "ladja": "ladja.png",
     "vesoljci": "vesoljci.png"
