+++ src/__main__.py
@@ -4,14 +4,15 @@ import os
 # adding directory to python path!
 sys.path.append(os.getcwd())
 
-from src.app.GUI import GUI
+from src.app.gui import Gui
 from src.settings import config as S
 
 S.init()
-app = GUI()
+app = Gui()
 app.init()
 
-while not app.konec():
+while True:
     app.narisi()
     app.vnos()
+    app.pocakaj()
 S.save()
