--- src/app/GUI.py
@@ -1,207 +1,0 @@
-import pygame_menu.widgets.core
-
-from src.app._app import App
-import sys
-import pygame
-import pygame_menu
-
-from src.domain.vesolje import Vesolje
-from src.utils import pot
-from src.settings import config as S
-
-
-class GUI(App):
-
-    def __init__(self):
-        self.sirina = S.CONFIG.pygame.dimenzija.sirina
-        self.visina = S.CONFIG.pygame.dimenzija.visina
-        pygame.init()
-
-        # NAJVEČJA DOVOLJENA RESOLUCIJA
-        self.info = pygame.display.Info()
-        self.surface = pygame.display.set_mode((self.sirina, self.visina), 0, 32)
-
-        # IZBIRA BARVNE TEME ZA MENIJE
-        self.theme = pygame_menu.themes.THEME_SOLARIZED
-        self.menu = pygame_menu.Menu(
-            S.JEZIK.aplikacija.naslov,
-            self.sirina,
-            self.visina,
-            theme=self.theme
-        )
-
-        self.config_menu = pygame_menu.Menu(
-            S.JEZIK.meni.konfiguracija,
-            S.CONFIG.pygame.dimenzija.sirina,
-            S.CONFIG.pygame.dimenzija.visina,
-            theme=self.theme
-        )
-
-        self.font = pygame.font.Font(
-            S.APP.font.tip,
-            S.APP.font.velikost
-        )
-
-        self.clock = pygame.time.Clock()
-
-        self.vnesi_ime_label = None
-
-        self.vesolje = None
-
-        self._init_meni()
-        self._init_config_meni()
-        self.menu.enable()
-        self.key_enabled = True
-
-    def _init_meni(self):
-        # INICIALIZACIJA GLAVNEGA MENIJA
-        self.vnesi_ime_label = self.menu.add.label(S.JEZIK.meni.vnesi_ime, background_color=S.APP.barve.bela)
-        self.ime_igralca = self.menu.add.text_input(S.JEZIK.meni.ime,
-                                                    onreturn=lambda ime: self._nastavi_ime(ime=ime))
-        self.menu.add.button(S.JEZIK.meni.igraj, self._zazeni_igro)
-        self.menu.add.button(S.JEZIK.meni.konfiguracija, self.config_menu)
-        self.menu.add.button(S.JEZIK.meni.izhod, pygame_menu.events.EXIT)
-
-    def _init_config_meni(self):
-        # VELIKOSTI ZASLONA
-        velikosti = [(S.JEZIK.meni_konfiguracija.zaslon.fullscreen, [self.info.current_w, self.info.current_h])]
-
-        for dim in S.CONFIG.pygame.dimenzije:
-            velikosti.append((f"{dim.sirina} x {dim.visina}", [dim.sirina, dim.visina]))
-        # IZBIRA JEZIKA
-        jeziki = []
-        for izbira in S.CONFIG.pygame.izbira_jezika:
-            jeziki.append((izbira.jezik, [izbira.jezik]))
-
-        self.config_menu.add.dropselect(S.JEZIK.meni_konfiguracija.zaslon.velikost, items=velikosti,
-                                        onchange=lambda _, velikost: self._nastavi_zaslon(velikost))
-        self.config_menu.add.dropselect(S.JEZIK.meni_konfiguracija.jezik, items=jeziki,
-                                        onchange=lambda _, tip: self._nastavi_jezik(tip))
-        self.config_menu.add.button(S.JEZIK.nazaj, pygame_menu.events.BACK)
-
-    def init(self):
-        self.vesolje = Vesolje()
-
-    def _mapiraj(self, x: float, y: float) -> tuple[int]:
-        return int(self.sirina * x), int(self.visina * y)
-
-    # TODO: da te dve funkciji implementiraš na način da ne
-    #  boš spreminjal nobene stvari iz domenskega prostora.
-    def _omejitev_pozicije_ladje(self):
-        sirina = 1 - self.vesolje.ladja.velikost_x
-        if self.vesolje.ladja.x >= sirina:
-            self.key_enabled = False
-        else:
-            self.key_enabled = True
-
-    # TODO: TO FUNKCIJO MI NI USPELO MODIFICIRAT
-    def _omejitev_pozicije_vesoljcev(self):
-        sirina = 1 - self.vesolje.velikost_vesoljca_x
-        if self.vesolje.vesoljci[-1].x >= sirina:
-            for j in range(len(self.vesolje.vesoljci)):
-                self.vesolje.vesoljci[j].y += self.vesolje.premik_vesoljca_navzdol
-                self.vesolje.vesoljci[j].hitrost *= -1
-
-    def narisi(self):
-        if self.menu.is_enabled():
-            self.menu.draw(self.surface)
-
-        else:
-            # MAPIRANJE POZICIJE IN VELIKOSTI LADJE
-            vl_x, vl_y = self._mapiraj(x=self.vesolje.ladja.velikost_x, y=self.vesolje.ladja.velikost_y)
-            v_x, v_y = self._mapiraj(x=self.vesolje.ladja.x, y=self.vesolje.ladja.y)
-
-            pygame.display.set_caption(S.JEZIK.aplikacija.naslov)
-            # ZAPOLNI DISPLAY Z BARVO, ČRNA
-            self.surface.fill(S.APP.barve.bela)
-            # SPREMENI VELIKOST
-            ladja = pygame.image.load(pot.media(S.APP.slike.ladja))
-
-            velikost_ladje = pygame.transform.scale(ladja, (vl_x, vl_y))
-
-            # NARIŠI LADJO
-            self.surface.blit(velikost_ladje, (v_x, v_y - vl_y))
-
-            vv_x, vv_y = self._mapiraj(x=self.vesolje.velikost_vesoljca_x,
-                                       y=self.vesolje.velikost_vesoljca_y)
-
-            vesoljci_slika = pygame.image.load(pot.media(S.APP.slike.vesoljci))
-            velikost_vesoljcev = pygame.transform.scale(vesoljci_slika, (vv_x, vv_y))
-            # MAPIRANJE IN RISANJE VESOLJCEV
-
-            for i in range(len(self.vesolje.vesoljci)):
-                ve_x, ve_y = self._mapiraj(x=self.vesolje.vesoljci[i].x, y=self.vesolje.vesoljci[i].y)
-                self.surface.blit(velikost_vesoljcev, (ve_x, ve_y + vv_y))
-                self.vesolje.vesoljci[i].premikanje()
-            self.vesolje.menjava_smeri_vesoljcev()
-            # self._omejitev_pozicije_vesoljcev()
-
-        pygame.display.update()
-
-    def vnos(self):
-        events = pygame.event.get()
-        if self.menu.is_enabled():
-            self.menu.update(events)
-            return
-        # SPREMLJAJ INPUTE V IGRI
-        for event in events:
-            if event.type == pygame.KEYDOWN:
-                if event.key == pygame.K_q:
-                    sys.exit()
-                elif event.key == pygame.K_SPACE:
-                    self.menu.enable()
-                # TODO: PRED MERGANJEM V MASTER ODSTRANITI EVENT ! TEST ONLY
-                # test odstrani vesoljce -> ENTER
-                elif event.key == pygame.K_RETURN:
-                    for vesoljec in self.vesolje.vesoljci:
-                        self.vesolje.vesoljci.remove(vesoljec)
-
-        keys = pygame.key.get_pressed()
-        if keys[pygame.K_a]:
-            self.vesolje.ladja.levo()
-            self.vesolje.omejitev_ladje()
-            self._omejitev_pozicije_ladje()
-
-        elif keys[pygame.K_d] and self.key_enabled:
-            self.vesolje.ladja.desno()
-            self.vesolje.omejitev_ladje()
-            self._omejitev_pozicije_ladje()
-
-        pygame.display.update()
-        self.clock.tick(S.APP.nastavitve.clock_tick)
-
-    def _izrisi_text(self, naslov: str, tekst: str, pozicija_x: int, pozicija_y: int):
-        # USTVARI TEXT PODLAGO
-        text_podlaga = self.font.render(f'{naslov} {tekst}', True, S.APP.barve.bela)
-        # NARIŠI TEXT NA CANVAS
-        self.surface.blit(text_podlaga, (pozicija_x, pozicija_y))
-
-    def _zazeni_igro(self):
-        self.menu.disable()
-
-    def _nastavi_ime(self, ime: str):
-        S.CONFIG.igralec.ime = ime
-        S.save()
-        self.ime_igralca.hide()
-        self.vnesi_ime_label.hide()
-
-    def _nastavi_zaslon(self, velikost_zaslona: list):
-        S.CONFIG.pygame.dimenzija.sirina = velikost_zaslona[0]
-        S.CONFIG.pygame.dimenzija.visina = velikost_zaslona[1]
-        S.save()
-
-    def _nastavi_jezik(self, jezik: list):
-        S.CONFIG.jezik = jezik[0]
-        S.save()
-
-    def konec(self) -> bool:
-        if len(self.vesolje.vesoljci) != 0:
-            for vesoljec in self.vesolje.vesoljci:
-                ve_x, ve_y = self._mapiraj(x=vesoljec.x, y=vesoljec.y)
-                vv_x, vv_y = self._mapiraj(x=self.vesolje.velikost_vesoljca_x,
-                                           y=self.vesolje.velikost_vesoljca_y)
-
-                if ve_y + vv_y >= self.visina:
-                    return True
-        else:
-            return True
