+++ src/app/_app.py
@@ -19,5 +19,9 @@ class App(ABC):
         pass
 
     @abstractmethod
+    def pocakaj(self):
+        pass
+
+    @abstractmethod
     def _mapiraj(self, x: float, y: float) -> tuple[int]:
         pass
