+++ src/domain/ladja.py
@@ -6,9 +6,7 @@ from src.domain.objekt import Objekt
 
 @dataclass
 class Ladja(Objekt):
-    velikost_x: float
-    velikost_y: float
-    hitrost: float
+    hitrost: float = 0.1
 
     def levo(self):
         self.x -= self.hitrost
