+++ src/domain/objekt.py
@@ -5,3 +5,8 @@ from dataclasses import dataclass
 class Objekt:
     x: float
     y: float
+    sirina:float
+    visina:float
+
+
+
