+++ src/domain/vector2D.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Vector2D:
+    x: float
+    y: float
